+++ src/tile_mate/stitcher.py
@@ -1,14 +1,18 @@
 from functools import lru_cache
 from pathlib import Path
 from typing import Optional
+import warnings
 
 import geopandas as gpd
 import rasterio
 from dem_stitcher.geojson_io import read_geojson_gzip
 from dem_stitcher.merge import merge_tile_datasets_within_extent
+from dem_stitcher.dateline import get_dateline_crossing
+from dem_stitcher.stitcher import _translate_one_tile_across_dateline
 from rasterio.errors import RasterioIOError
 from rasterio.env import Env
 from shapely.geometry import box
+import pandas as pd
 
 from .exceptions import NoTileCoverage, TilesetNotSupported
 from .tile_model import TILE_SCHEMA
@@ -128,6 +132,7 @@ def get_tile_data(
 
     if df_tiles.empty:
         raise NoTileCoverage(f'{tile_key} has no global tiles with the parameters provided')
+
     return df_tiles
 
 
@@ -152,9 +157,22 @@ def update_hansen_landsat_mosaic_url(url: str, year: int):
 
 
 def get_urls_from_tile_df(extent: list[float], df_tiles: gpd.GeoDataFrame) -> list[str]:
+    df_tiles_all = df_tiles.copy()
+    crossing = get_dateline_crossing(extent)
+    if crossing:
+        warnings.warn(
+            'Getting tiles across dateline on the opposite hemisphere; '
+            f'The source tiles will be {-2 * crossing} deg along the '
+            'longitudinal axis from the extent requested',
+            category=UserWarning,
+        )
+        df_tiles_all_translated = df_tiles_all.copy()
+        x_translation = 2 * crossing
+        df_tiles_all_translated.geometry = df_tiles_all.geometry.translate(xoff=x_translation)
+        df_tiles_all = pd.concat([df_tiles_all, df_tiles_all_translated], axis=0).reset_index(drop=True)
     bbox = box(*extent)
-    ind_inter = df_tiles.geometry.intersects(bbox)
-    df_subset = df_tiles[ind_inter].reset_index(drop=True)
+    ind_inter = df_tiles_all.geometry.intersects(bbox)
+    df_subset = df_tiles_all[ind_inter].reset_index(drop=True)
     urls = df_subset.url.tolist()
     if not urls:
         raise NoTileCoverage('There are no tiles over the requested area')
@@ -218,13 +236,21 @@ def get_raster_from_tiles(
 
     urls_subset = get_urls_from_tile_df(extent, df_tiles)
 
+    datasets = [rasterio.open(url) for url in urls_subset]
+    crossing = get_dateline_crossing(extent)
+    if crossing:
+        zipped_data = list(map(lambda ds: _translate_one_tile_across_dateline(ds, crossing), datasets))
+        memory_files, datasets = zip(*zipped_data)
+
     if tile_shortname == 'umd_ocean_mask':
         env = Env(GS_NO_SIGN_REQUEST='YES')
     else:
         env = Env()  # default environment
 
     with env:
-        X_merged, p_merged = merge_tile_datasets_within_extent(urls_subset, extent)
+        X_merged, p_merged = merge_tile_datasets_within_extent(datasets, extent)
+    if crossing:
+        list(map(lambda mf: mf.close(), memory_files))
 
     # Are stored in the profile for provenance
     p_merged.update(**tile_metadata)